// src/__generated__/branch-query.ts
/* eslint-disable */
/* This is an autogenerated file. Do not edit this file directly! */
export type BranchQuery = {
    repository: ({
        ref: ({
            name: string;
        }) | null;
    }) | null;
};
export type BranchQueryVariables = {
    name: string;
    owner: string;
    repo: string;
};